refactor : 콘솔용 어플리케이션, 컨트롤러임을 나타내도록 클래스 이름 변경 및 멤버변수 선언 위치 수정
ConsoleChessController를 동시에 여러 개 만들어 사용하게 되면
static으로 선언한 멤버변수 board의 상태 변화가 공유되므로 멤버변수 board를 인스턴스 변수로 수정
