--- src/main/java/chess/Application.java
@@ -1,9 +1,0 @@
-package chess;
-
-public class Application {
-
-    public static void main(String[] args) {
-        ChessController chessController = new ChessController();
-        chessController.run();
-    }
-}
